+++ addon/components/native-select-box/group.js
@@ -1,6 +1,6 @@
 import Component from '@ember/component';
-import Nameable from '../../../mixins/general/nameable';
-import layout from '../../../templates/components/select-box/native/group';
+import Nameable from '../../mixins/general/nameable';
+import layout from '../../templates/components/native-select-box/group';
 
 const mixins = [
   Nameable