+++ tests/integration/components/native-select-box/option-test.js
@@ -4,13 +4,13 @@ import { render, find, findAll, settled } from '@ember/test-helpers';
 import hbs from 'htmlbars-inline-precompile';
 import { defer } from 'rsvp';
 
-module('select-box/native/option', function(hooks) {
+module('native-select-box/option', function(hooks) {
   setupRenderingTest(hooks);
 
   test('it renders', async function(assert) {
     assert.expect(1);
 
-    await render(hbs `{{select-box/native/option}}`);
+    await render(hbs `{{native-select-box/option}}`);
 
     assert.equal(findAll('option.select-box-option').length, 1,
       'renders with correct class name and tag');
@@ -19,7 +19,7 @@ module('select-box/native/option', function(hooks) {
   test('class prefix', async function(assert) {
     assert.expect(1);
 
-    await render(hbs `{{select-box/native/option class-prefix="foo"}}`);
+    await render(hbs `{{native-select-box/option class-prefix="foo"}}`);
 
     assert.equal(findAll('.foo-option').length, 1,
       'can override the class prefix');
@@ -28,7 +28,7 @@ module('select-box/native/option', function(hooks) {
   test('title', async function(assert) {
     assert.expect(1);
 
-    await render(hbs `{{select-box/native/option title="Foo"}}`);
+    await render(hbs `{{native-select-box/option title="Foo"}}`);
 
     assert.equal(find('.select-box-option').getAttribute('title'), 'Foo',
       'a native select box option can have a title attribute');
@@ -39,7 +39,7 @@ module('select-box/native/option', function(hooks) {
 
     this.set('optionDisabled', true);
 
-    await render(hbs `{{select-box/native/option disabled=this.optionDisabled}}`);
+    await render(hbs `{{native-select-box/option disabled=this.optionDisabled}}`);
 
     assert.ok(find('.select-box-option').disabled,
       'a native select box option can be disabled');
@@ -55,7 +55,7 @@ module('select-box/native/option', function(hooks) {
 
     this.set('myValue', 123);
 
-    await render(hbs `{{select-box/native/option value=this.myValue}}`);
+    await render(hbs `{{native-select-box/option value=this.myValue}}`);
 
     assert.strictEqual(find('.select-box-option').getAttribute('value'), '123',
       'the specified value is set as an HTML attribute');
@@ -74,9 +74,9 @@ module('select-box/native/option', function(hooks) {
     this.set('myValue', deferred.promise);
 
     await render(hbs `
-      {{#select-box/native/option value=this.myValue as |o|}}
+      {{#native-select-box/option value=this.myValue as |o|}}
         {{~myValue}}: {{o.value~}}
-      {{/select-box/native/option}}
+      {{/native-select-box/option}}
     `);
 
     assert.equal(find('.select-box-option').textContent, '[object Object]: [object Object]',
@@ -100,9 +100,9 @@ module('select-box/native/option', function(hooks) {
     assert.expect(1);
 
     await render(hbs `
-      {{#select-box/native/option~}}
+      {{#native-select-box/option~}}
         Foo
-      {{~/select-box/native/option}}
+      {{~/native-select-box/option}}
     `);
 
     assert.strictEqual(find('.select-box-option').textContent, 'Foo',
@@ -118,13 +118,13 @@ module('select-box/native/option', function(hooks) {
     ]);
 
     await render(hbs`
-      {{#select-box/native value="foo" as |sb|}}
+      {{#native-select-box value="foo" as |sb|}}
         {{#each this.items as |item|}}
           {{#sb.option value=item as |o|~}}
             {{o.index}}={{o.value}}
           {{~/sb.option}}
         {{/each}}
-      {{/select-box/native}}
+      {{/native-select-box}}
     `);
 
     assert.ok(